Simplify Setup of Acima Payment Method
Previously we requested users to input the iframe url that they would be using
for their application. However this URL only has 2 options: sandbox and production
and it's the same accross all vendors. To simplify setup and avoid errors when
copy-pasting the url, we've removed this input and instead will generate a url
based on whether the payment is marked as 'test_env: true' or not.