Handle order completion for apps without confirmation step
In the previous variant we still expected the gem user to handle the step of completing
an order and redirecting to the correct URL in their own app with a dedicated endpoint.
With this commit we remove that expectation as we can handle the order completion and
redirection ourselves provided we get the correct data attributes.