Rework logic when custom url is provided
The need of adding a custom URL came from removing the confirmation step in the
order flow. However, we found out that with that step removed the PATCH
CheckoutsController#update would return a 422 error as it couldn't advance to the
next state.

Considering the existing implementation we decided to add a POST PaymentsController
#create step assuming that the user will run an @order.complete on their side.
This however isn't a good implementation and should be modified in the future.